requirements.txt
SpeechRecognition
spacy
ytmusicapi
python-vlc
pafy
youtube-dl
pyttsx3
PyAudio